Merged string support + gradle-build + sv-comp compatible script (#105)
* Z3str3 string solving support (GSoC 2022 submission) (#78)

* updated .gitignore, uncached .jpf files

* initial z3str3 files from v2 files

* init v3 files working as proc

* z3String3Processor cleanup

* failing jni calls

* JNI calls working, not parsing output yet

* parsing output back to SPF OK, crude...

* close context

* boolean z3str3 parms from jpf -> z3str3processor

* stripping out (check-sat) (get-model) from query

* exception handling, unsat get-model prevention

* Z3String3processor cleanup

* removed proc interface from z3string3processor

* working BSU solver integration

* z3str3 and general doc

* post gh failure repair

* more gh fail repairs

* more post gh fail repairs

* rep .gitignore

* hamcrest jar repl

* Add more examples for various String operations

* Add support to String.isEmpty()

Handle isEmpty() using existing support of String.length().
e.g. if string length is 0 then isEmpty() return true.

* Add SV-COMP examples

* Ignore build/ folder

* Update examples and add new config files

- Add test for valueOf(char)
- Add preliminary config file for CRIME, MasterMind and StringSearch05
- Use z3 and z3str3 as numeric and string solvers, respectively

* Fix syntax error for DelimSearch config file

* Update z3 library to version 4.8.17 for windows

* Minor fixes according to Java Range branch

Fixes in terms of df100ff2be48d2e1be77c5654bee4c9bfbe34ad3 and 6863780ba97bb3fbe06abaa83be15ee4d9e7e309

* Fix type inconsistency for String.CharAt() modeling

CharAt() was mapped to str.at in stmlib. The problem is that Java CharAt() return an char (integer in Java bytecode) while str.at return a string.
To fix this inconsistency, conver the output string to an integer using str.to_code

* Add preliminary support for valueOf(char)

* Check whether strings exist in query result from Z3str3 before parsing

Only parse the result if strings exist. Ignore the result otherwise.
This fix is a simple workaround for a StringIndexOutOfBoundsException

* Limit input integer size for strings example Tricky

To avoid recursion for too many times, limit the input so that k for bar(int n, int k, String op) is non negative.
Also turn on debug mode for this example.

* Improve valueOf(char) modeling

Since normal char are integers with range 0 - 65535 while charAt() results are modeled as strings, handle them seperately in valueOf(char)

* Add example calling charAt() on empty string

Note that this example will raise an exception from JPF

* Revert "Fix type inconsistency for String.CharAt() modeling"

This reverts commit 04185dadc2eeb60183542fbd43eac21d47cc336e.

* Handle (CharAt() & 65535) case

This case happens when casting the result of CharAt() to integer and then cast the integer back to char.
For this case, the bitwise and is removed when modeling to Z3 SMT formula since it is unnecessary.

For a more general case (bitwise AND between two arbitrary integers), both integers are converted to bitvectors, AND with bvand and the result is converted back to integer.
TODO: this approach is impractically slow, find a more efficient workaround for integer AND modeling.

* Propose change on Tricky in example/strings

* Set up CRIME to run outside of SPF

* Set up SPF execution for MasterMind in examples/strings

* Add support to ValueOf(int)

* Update com.microsoft.z3 to Z3 version 4.8.17

* Add Junit 4 library

Required when running tests in src/tests/

* Comment out all tests in E5 run config

Tests are all commented out by default.
Uncomment to run.

* Fix inconsistency between string tests and Z3 interface

TestSymString is inconsist with current Z3str3 support. Fixes as below:
- Create Z3 compatible version of TestSymString.Test1() as a example
- Copy solver returned solutions to spc.solution

* Add partial support to Integer.parseInt()

Add support to isInteger and notinteger, which is required by parseInt()

* Add simple testing example for trim()

* Fix operands order for startdWith() and endsWith()

* GSoC 2022 document 1st draft

* GSoC 2022 document revision

* Preliminary support to trim()

This implementation only work when this is only one unique occurence of trim() in the same solver query

* Change MasterMind solver to z3

* ParseInt() exception handling

SPF previously throw a exception of itself if ParseInt() end with a exception.
Instead, raise a simulated NumberFormatException inside SPF and continue exploring other paths.

* Add pattern matching for valueOf(char)

---------

Co-authored-by: Marlin Roberts <marlinroberts@u.boisestate.edu>
Co-authored-by: Marlin Roberts <51375407+marlinroberts21@users.noreply.github.com>

* adding sv-comp script and creating a default site.properties for sv-comp, since creation of files is no longer allowed

---------

Co-authored-by: Qiuchen Yan <yanxx297@umn.edu>
Co-authored-by: Marlin Roberts <marlinroberts@u.boisestate.edu>
Co-authored-by: Marlin Roberts <51375407+marlinroberts21@users.noreply.github.com>